refactor(step 2): extract FlowLayout into its own file (#1976)
